third_party/blink/web_tests/http/tests/devtools/elements/styles-1/filter-matched-styles-expected.txt

Verifies that filtering in StylesSidebarPane works as expected.


Running: selectInitialNode

Running: testSimpleFiltering
Filtering styles by: padding
.mydiv {
   border:>1px solid black
F  padding:v10px 10px 10px 10px
F      padding-top: 10px
F      padding-right: 10px
F      padding-bottom: 10px
F      padding-left: 10px
}

Running: testLonghandsAreAutoExpanded
Filtering styles by: -top
element.style {
   margin:v1px
F      margin-top: 1px
       margin-right: 1px
       margin-bottom: 1px
       margin-left: 1px
}
.mydiv {
   border:v1px solid black
F      border-top-width: 1px
       border-right-width: 1px
       border-bottom-width: 1px
       border-left-width: 1px
F      border-top-style: solid
       border-right-style: solid
       border-bottom-style: solid
       border-left-style: solid
F      border-top-color: black
       border-right-color: black
       border-bottom-color: black
       border-left-color: black
       border-image-source: initial
       border-image-slice: initial
       border-image-width: initial
       border-image-outset: initial
       border-image-repeat: initial
   padding:v10px 10px 10px 10px
F      padding-top: 10px
       padding-right: 10px
       padding-bottom: 10px
       padding-left: 10px
}

Running: testAutoExpandedLonghandsAreCollapsed
Filtering styles by: null
element.style {
   margin:>1px
}
#inspected {
   border-size: 2px
}
.mydiv {
   border:>1px solid black
   padding:>10px 10px 10px 10px
}
div {
   display: block
}
